Split the style into one module per layer family
Layer structure moves out of the 850-line index into per-layer modules —
depth-shading (with the relief ramp), depth-areas (with the isobath
ladders and the safety-contour snap both it and the contours share),
hillshade, contours, soundings, coverage — with appearance staying in
flavor.ts and index.ts reduced to composition: sources(), layers(),
style(), applyState(), readDepth(). The public API is unchanged, and
every expression moved verbatim.

Each layer module also declares WHICH of its properties depend on the
mariner settings (unit/safety/shading/hillshade), and applyState walks
those declarations generically. The old hand-maintained per-layer switch
silently skipped any state-dependent property it didn't know about — a
layer gaining one now declares it in the same file, same diff. The
generic walk emits the identical call set to the switch it replaces.

Relative imports carry .js extensions so the built package resolves
under pure Node ESM, not just bundlers, and package.json ships every
module for the development condition that serves raw TypeScript.
