+++ style/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@openwaters/seascape",
   "version": "0.3.0",
-  "description": "MapLibre GL style for Open Waters bathymetry tiles — depth shading, contours, soundings, depth areas",
+  "description": "MapLibre GL style for Open Waters bathymetry tiles \u2014 depth shading, contours, soundings, depth areas",
   "license": "BSD-3-Clause",
   "type": "module",
   "main": "dist/index.js",
@@ -15,8 +15,8 @@
   },
   "files": [
     "dist",
-    "index.ts",
-    "README.md"
+    "*.ts",
+    "!*.test.ts"
   ],
   "repository": {
     "type": "git",
